refactor(components): rename a StyledFlex filename to DodamFlex
