+++ packages/components/src/layout/Flex/index.ts
@@ -1,2 +1,1 @@
-export * from "./FlexLayout";
-export * from "./StyledFlex";
+export * from "./DodamFlex";
